+++ server/statistics/store.go
@@ -265,43 +265,3 @@ func (r *RollingStoreStats) GetInstantLoad(k StoreStatKind) float64 {
 	}
 	return 0
 }
-
-// SlowStoresStats is a cached statistics for the slow store.
-type SlowStoresStats struct {
-	syncutil.RWMutex
-	slowStores map[uint64]struct{}
-}
-
-// NewSlowStoresStats creates a new slowStoresStats cache.
-func NewSlowStoresStats() *SlowStoresStats {
-	return &SlowStoresStats{
-		slowStores: make(map[uint64]struct{}),
-	}
-}
-
-// RemoveSlowStoreStatus removes SlowStoreStats with a given store ID.
-func (s *SlowStoresStats) RemoveSlowStoreStatus(storeID uint64) {
-	s.Lock()
-	defer s.Unlock()
-	delete(s.slowStores, storeID)
-}
-
-// ObserveSlowStoreStatus updates SlowStoreStats with a given store ID.
-func (s *SlowStoresStats) ObserveSlowStoreStatus(storeID uint64, isSlow bool) {
-	s.Lock()
-	defer s.Unlock()
-	// If the given store was slow, this store should be recorded. Otherwise,
-	// this store should be removed from the recording list.
-	if _, ok := s.slowStores[storeID]; !ok && isSlow {
-		s.slowStores[storeID] = struct{}{}
-	} else {
-		delete(s.slowStores, storeID)
-	}
-}
-
-// ExistsSlowStores returns whether there exists slow stores in this cluster.
-func (s *SlowStoresStats) ExistsSlowStores() bool {
-	s.RLock()
-	defer s.RUnlock()
-	return len(s.slowStores) > 0
-}
